removed useless code, changed the write mechanism to use mutx instead of channels

+++ db/db.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"path/filepath"
 	"strconv"
-	"sync/atomic"
+	"sync"
 	"time"
 
 	"github.com/alash3al/redix/db/driver"
@@ -17,17 +17,19 @@ var (
 
 // DB represents datastore
 type DB struct {
-	db       driver.Interface
-	name     string
-	driver   string
-	writes   atomic.Value
-	counters atomic.Value
-	tmp      chan driver.KeyValue
+	db           driver.Interface
+	name         string
+	driver       string
+	writes       chan driver.KeyValue
+	buffer       []driver.KeyValue
+	useAsyncMode bool
+	l            *sync.RWMutex
 }
 
 // Open a database
 func Open(drivername, dirname string, dbname string, opts map[string]interface{}) (*DB, error) {
 	dbkey := fmt.Sprintf("%s_%s", drivername, dbname)
+
 	if dbInterface, loaded := databases.Load(dbkey); loaded {
 		return dbInterface.(*DB), nil
 	}
@@ -43,50 +45,33 @@ func Open(drivername, dirname string, dbname string, opts map[string]interface{}
 	}
 
 	db := &DB{
-		db:       driverInstance,
-		name:     dbname,
-		driver:   drivername,
-		writes:   atomic.Value{},
-		counters: atomic.Value{},
+		db:           driverInstance,
+		name:         dbname,
+		driver:       drivername,
+		writes:       make(chan driver.KeyValue, 10000),
+		buffer:       []driver.KeyValue{},
+		l:            &sync.RWMutex{},
+		useAsyncMode: opts["async"].(bool),
 	}
 
-	db.writes.Store([]driver.KeyValue{})
+	databases.Store(dbkey, db)
 
 	go (func() {
-		tk := time.NewTicker(100 * time.Millisecond)
-		defer tk.Stop()
+		for {
+			time.Sleep(time.Second * 1)
 
-		for range tk.C {
-			snapshot := db.writes.Load()
-			db.writes.Store([]driver.KeyValue{})
-			pairs := snapshot.([]driver.KeyValue)
+			db.l.Lock()
+			newBuffer := make([]driver.KeyValue, len(db.buffer))
+			copy(newBuffer, db.buffer)
+			db.buffer = db.buffer[:0]
 
-			for _, pair := range pairs {
-				if pair.Value == nil {
-					db.delete(pair.Key)
-				} else {
-					db.put(pair.Key, pair.Value, pair.TTL)
-				}
+			db.l.Unlock()
+
+			for _, pair := range newBuffer {
+				db.putSync(pair.Key, pair.Value, pair.TTL)
 			}
 		}
 	})()
-
-	counters := map[string]int64{}
-	db.Scan(driver.ScanOpts{
-		Prefix:        countersPrefix,
-		IncludeOffset: true,
-		Scanner: func(k, v []byte) bool {
-			k = k[len(countersPrefix)+1:]
-
-			val, _ := strconv.ParseInt(string(v), 10, 64)
-			counters[string(k)] = val
-
-			return true
-		},
-	})
-	db.counters.Store(counters)
-
-	databases.Store(dbkey, db)
 
 	return db, nil
 }
@@ -107,49 +92,13 @@ func (db DB) Name() string {
 	return db.name
 }
 
-// PutAsync puts new document into the storage [async]
-func (db *DB) PutAsync(key []byte, value []byte, ttl int) error {
-	var newVals []driver.KeyValue
-
-	pair := driver.KeyValue{Key: key, Value: value, TTL: ttl}
-	oldVals := db.writes.Load().([]driver.KeyValue)
-
-	copy(newVals, oldVals)
-
-	newVals = append(newVals, pair)
-
-	db.writes.Store(newVals)
-
-	return nil
-}
-
-// Put puts new document into the storage [sync]
-func (db DB) Put(key []byte, value []byte, ttl int) error {
-	return db.put(key, value, ttl)
-}
-
-// // Batch a bulk pairs writer
-// func (db DB) Batch(pairs []driver.KeyValue) error {
-// 	return db.db.Batch(pairs)
-// }
-
-// Incr increments a key
-func (db *DB) Incr(key []byte, delta int) (int64, error) {
-	oldCounters := db.counters.Load().(map[string]int64)
-	newCounters := map[string]int64{}
-
-	for k, v := range oldCounters {
-		newCounters[k] = v
+// Put writes a key - value pair into the database
+func (db DB) Put(key, value []byte, ttl int) error {
+	if db.useAsyncMode {
+		return db.putAsync(key, value, ttl)
 	}
 
-	key = append(countersPrefix, key...)
-
-	newCounters[string(key)] += int64(delta)
-	newVal := newCounters[string(key)]
-
-	db.counters.Store(newCounters)
-
-	return newVal, db.PutAsync(key, []byte(strconv.FormatInt(newVal, 10)), 0)
+	return db.putSync(key, value, ttl)
 }
 
 // Get fetches a document using its primary key
@@ -167,7 +116,7 @@ func (db DB) Get(key []byte) ([]byte, error) {
 
 	expires, _ := strconv.ParseInt(string(expirable), 10, 64)
 	if expires > 0 && (time.Now().UnixNano() >= expires) {
-		db.DeleteAsync(key)
+		db.putAsync(key, nil, 0)
 		return nil, nil
 	}
 
@@ -182,18 +131,6 @@ func (db DB) Has(key []byte) (bool, error) {
 // Delete remove a key from the database [sync]
 func (db DB) Delete(key []byte) error {
 	return db.db.Delete(key)
-}
-
-// DeleteAsync remove a key from the database [async]
-func (db DB) DeleteAsync(key []byte) error {
-	pair := driver.KeyValue{Key: key, Value: nil}
-
-	snapshot := db.writes.Load()
-	vals := snapshot.([]driver.KeyValue)
-	vals = append(vals, pair)
-	db.writes.Store(vals)
-
-	return nil
 }
 
 // Scan scans the db
@@ -211,17 +148,25 @@ func (db DB) Close() error {
 	return db.db.Close()
 }
 
-// Delete remove a key from the database [async]
-func (db DB) delete(key []byte) error {
-	return db.db.Delete(key)
-}
+// Put puts new document into the storage [sync]
+func (db DB) putSync(key []byte, value []byte, ttl int) error {
+	if nil == value {
+		return db.db.Delete(key)
+	}
 
-// put puts new document into the storage [sync]
-func (db DB) put(key []byte, value []byte, ttl int) error {
 	if ttl > 0 {
 		expires := time.Now().Add(time.Duration(ttl) * time.Millisecond).UnixNano()
 		db.db.Put(append(expirablePrefix, key...), []byte(strconv.FormatInt(expires, 10)))
 	}
 
 	return db.db.Put(key, value)
+}
+
+// PutAsync puts new document into the storage [async]
+func (db *DB) putAsync(key []byte, value []byte, ttl int) error {
+	db.l.Lock()
+	db.buffer = append(db.buffer, driver.KeyValue{Key: key, Value: value, TTL: ttl})
+	db.l.Unlock()
+
+	return nil
 }
